+++ src/PdfTools.h
@@ -9,3 +9,4 @@ void ShowPdfCompressDialog(MainWindow* win);
 void ShowPdfDecompressDialog(MainWindow* win);
 void ShowPdfDeletePageDialog(MainWindow* win);
 void ShowPdfExtractPagesDialog(MainWindow* win);
+void ShowPdfEncryptDialog(MainWindow* win);