+++ src/SumatraPDF.cpp
@@ -6898,6 +6898,10 @@ static LRESULT FrameOnCommand(MainWindow* win, HWND hwnd, UINT msg, WPARAM wp, L
             ShowPdfExtractPagesDialog(win);
             break;
 
+        case CmdPdfEncrypt:
+            ShowPdfEncryptDialog(win);
+            break;
+
         case CmdDocumentExtractText:
             ShowPdfExtractTextDialog(win);
             break;